Remove outline module
the outline and config modules are really for the same task, so I
believe that they should belong in the same module, this combines the
two of them.
